--- backend/package.json
@@ -1,17 +1,0 @@
-{
-  "name": "backend",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node app.js"
-  },
-  "author": "mohammed zaariy",
-  "license": "ISC",
-  "dependencies": {
-    "cros": "^1.0.1",
-    "express": "^4.18.0",
-    "nodemon": "^2.0.15"
-  }
-}
