--- src/CSS/faq.css
@@ -1,10 +1,0 @@
-.faq {
-    height: 50vh;
-}
-.faq h1 {
-    font-weight: 600;
-    width: 100%;
-    text-align: center;
-    margin-top: 20px;
-    
-}
