--- src/components/Faq.jsx
@@ -1,14 +1,0 @@
-import React  from "react";
-import '../CSS/faq.css'
-function Faq() {
-
-    return (
-        <div className="faq">
-            <div className="container">
-                <h1>FAQ</h1>
-            </div>
-        </div>
-    )
-}
-
-export default Faq ;
