+++ src/pages/FAQ/index.jsx
@@ -1,0 +1,14 @@
+import React from "react";
+import './style/main.css'
+function Faq() {
+
+    return (
+        <div className="faq">
+            <div className="container">
+                <h1>FAQ</h1>
+            </div>
+        </div>
+    )
+}
+
+export default Faq;
