+++ src/pages/FAQ/style/main.css
@@ -1,0 +1,11 @@
+.faq {
+    height: 50vh;
+}
+
+.faq h1 {
+    font-weight: 600;
+    width: 100%;
+    text-align: center;
+    margin-top: 20px;
+
+}
